Add support to edit and delete user data and prescriptions. Add new tests
Now, user phenotype data can be edited, and all the user-related data (phenotype, prescriptions and test) can be deleted from the DB. Prescriptions from a patient can be edited too.

Also, the form to add new test it's ready (but the data can not be saved yet), with help for each question of the tests.