Let you send messages to patients
Communication section is now implemented, allowing users to send messages to patients. That messages can be sent inmediatly or can be scheduled to be sent to certain hours in a given date interval.  Also, you can edit and delete messages, and filtering them by date.

Communication section is used as Prescriptions.