Add viz view, add prescription dataset modal and API methods
Now the "Visualización" (dataviz) option is available in the navbar menu. Now we can add prescription datasets from patients and plot it using dimple.js. The datasets are obtained via AJAX requests to an API implemented in app.js file.

The view has a modal form to search for patients and add them to the patient dataset's list to be ploted. 

package.json is updated to include the connect-rest dependency. Also, we include some handlebars templates to render the search results obtained from the modal in the visualization view.