Fix some issues on dataviz chart
The chart plotting waits until the datasets are fully loaded  (using jQuery.when function) to draw itself. All the logic of the chart is in "dimpleChart.js" file. 
Also, the deleteChart() procedure works fine and lets delete a chart before plotting a new one.

FInally, in main.handlebars layout, we set default column grid to 12 instead the 9 that were before in that views which not contain patients' data. That's the reason because the dataviz charts is now bigger!